[Refactor] QuizViewController View 로직 분리작업
