+++ ChattingApp/ChattingApp/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  ChattingApp
+//
+//  Created by 김도현 on 2023/10/22.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
