+++ oewoboka/oewoboka/Scenes/QuizPage/QuizViewController.swift
@@ -10,57 +10,16 @@ import UIKit
 import SnapKit
 
 final class QuizViewController: UIViewController {
-    
-    private lazy var frontCardView: CardView = CardView(viewModel: viewModel)
-    private lazy var backgroundCardView: CardView = CardView(viewModel: viewModel)
-    private lazy var quizControlView: QuizControlStackView = QuizControlStackView(buttonSize: CGSize(width: 52.0, height: 52.0), viewModel: viewModel)
-    private lazy var dictationStackView: UIStackView = {
-        let stackView = UIStackView(arrangedSubviews: [
-            dictationTextField
-        ])
-        stackView.isLayoutMarginsRelativeArrangement = true
-        stackView.layoutMargins = UIEdgeInsets(top: 8, left: 0, bottom: 8, right: 0)
-        stackView.axis = .horizontal
-        stackView.alignment = .fill
-        stackView.distribution = .fill
-        stackView.spacing = 12
-        return stackView
-    }()
-    private lazy var dictationTextField: UITextField = {
-        let textField = UITextField()
-        textField.placeholder = viewModel.isMeanDictation ? "의미를 작성한 뒤 엔터를 눌러주세요" : "단어를 작성한 뒤 엔터를 눌러주세요"
-        textField.font = Typography.title3.font
-        textField.textColor = .black
-        textField.borderStyle = .roundedRect
-        return textField
-    }()
-    private let countDownLabel: UILabel = {
-        let label = UILabel()
-        label.font = Typography.title3.font
-        label.textColor = .black
-        label.text = "00시00분00초"
-        return label
-    }()
-    private lazy var timerSettingView: TimerSettingView = TimerSettingView(viewModel: viewModel)
-    private var timer = Timer()
-    private var countDownTime: Int = 0 {
-        didSet {
-            timerSettingView.isHidden = true
-            let minute = 60
-            if countDownTime <= minute {
-                countDownLabel.textColor = .systemRed
-            }
-            countDownLabel.text = countDownTime.stringFromTime()
-        }
-    }
-    private let margin: CGFloat = 24
-    private var currentIndex: Int = 0
-    private var bottomConstraint: Constraint?
 
+    private var timer = Timer()
+    private let quizMainView: QuizMainView
     private let viewModel: QuizViewModel
+    
+    private let margin: CGFloat = 24
     
     init(quizData: QuizSettingData) {
         self.viewModel = QuizViewModel(quizData: quizData)
+        self.quizMainView = QuizMainView(viewModel: viewModel)
         super.init(nibName: nil, bundle: nil)
         view.backgroundColor = .white
     }
@@ -89,33 +48,25 @@ final class QuizViewController: UIViewController {
 private extension QuizViewController {
     func setup() {
         addViews()
-        navigationSetup()
-        initCardView()
-        cardViewSetup()
+        setupAutoLayout()
+        setupNavigation()
+        setupCardView()
+        addCardViewGesture()
         textFieldSetup()
-        autoLayoutSetup()
         bind()
     }
     
     func addViews() {
-        view.addSubview(backgroundCardView)
-        view.addSubview(frontCardView)
-        if viewModel.isWordCard {
-            view.addSubview(quizControlView)
-        } else {
-            view.addSubview(dictationStackView)
-            if viewModel.isTestType {
-                view.addSubview(timerSettingView)
-                view.addSubview(countDownLabel)
-                timerSettingView.backgroundColor = .white
-                timerSettingView.snp.makeConstraints { make in
-                    make.center.equalToSuperview()
-                }
-            }
+        view.addSubview(quizMainView)
+    }
+    
+    func setupAutoLayout() {
+        quizMainView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
         }
     }
     
-    func navigationSetup() {
+    func setupNavigation() {
         let leftBarButtonItem = UIBarButtonItem(image: UIImage(systemName: "chevron.left"), style: .done, target: self, action: #selector(popViewController))
         leftBarButtonItem.tintColor = .black
         navigationItem.leftBarButtonItem = leftBarButtonItem
@@ -126,110 +77,38 @@ private extension QuizViewController {
         dismiss(animated: true)
     }
     
-    func initCardView() {
+    func addCardViewGesture() {
         
         if viewModel.isWordCard {
             let panGesture = UIPanGestureRecognizer(target: self, action: #selector(cardMove(sender:)))
-            frontCardView.addGestureRecognizer(panGesture)
+            quizMainView.frontCardView.addGestureRecognizer(panGesture)
         }
         let tapGesture = UITapGestureRecognizer(target: self, action: #selector(keyboardDismiss(sender:)))
-        frontCardView.isUserInteractionEnabled = true
-        frontCardView.addGestureRecognizer(tapGesture)
-        
-        frontCardView.layer.borderWidth = 1
-        frontCardView.layer.borderColor = UIColor.black.cgColor
-        frontCardView.layer.cornerRadius = 10
-        frontCardView.layer.masksToBounds = true
-        
-        backgroundCardView.layer.borderWidth = 1
-        backgroundCardView.layer.borderColor = UIColor.black.cgColor
-        backgroundCardView.layer.cornerRadius = 10
-        backgroundCardView.layer.masksToBounds = true
+        quizMainView.frontCardView.isUserInteractionEnabled = true
+        quizMainView.frontCardView.addGestureRecognizer(tapGesture)
     }
     
     @objc func keyboardDismiss(sender: UITapGestureRecognizer) {
         view.endEditing(true)
     }
-    
-    func cardViewSetup() {
-        if viewModel.isComplete {
-            navigationPushQuizCompletePage()
-            return
-        }
-        
-        frontCardView.englishLabel.text = viewModel.currentEnglishLabelText
-        frontCardView.koreaLabel.text = viewModel.currentKoreaLabelText
-        frontCardView.wordCountLabel.text = viewModel.progressCountText
-        
-        frontCardView.englishLabel.isHidden = viewModel.isWordLabelHidden
-        frontCardView.koreaLabel.isHidden = viewModel.isMeanLabelHidden
 
-        frontCardView.bookMarkButton.isSelected = viewModel.currentWord.isBookmark
-        
-        guard let nextWord = viewModel.nextWord else {
-            backgroundCardView.isHidden = true
-            return
-        }
-        
-        backgroundCardView.wordCountLabel.text = viewModel.nextProgressCountText
-        backgroundCardView.englishLabel.text = viewModel.nextEnglishText
-        backgroundCardView.koreaLabel.text = viewModel.nextKoreaText
-        backgroundCardView.bookMarkButton.isSelected = nextWord.isBookmark
-    }
-    
-    func autoLayoutSetup() {
-        let safeArea = view.safeAreaLayoutGuide
-        var bottomViewHeight = 0.0
-        var cardViewBottomConstraint: Constraint? = nil
-        
-        if viewModel.isTestType {
-            countDownLabel.snp.makeConstraints { make in
-                make.top.left.equalTo(safeArea).inset(margin)
-            }
-        }
-        
-        frontCardView.snp.makeConstraints { make in
-            if viewModel.isTestType {
-                make.top.equalTo(countDownLabel.snp.bottom).offset(12)
-            } else {
-                make.top.equalTo(safeArea).inset(margin)
-            }
-            make.left.right.equalTo(safeArea).inset(margin)
-            cardViewBottomConstraint = make.bottom.equalTo(safeArea).constraint
-        }
-        backgroundCardView.snp.makeConstraints { make in
-            make.edges.equalTo(frontCardView)
-        }
-        
-        if viewModel.isWordCard {
-            quizControlView.snp.makeConstraints { make in
-                make.top.equalTo(frontCardView.snp.bottom).offset(margin)
-                make.height.equalTo(52)
-                make.left.right.bottom.equalTo(safeArea).inset(margin)
-            }
-            quizControlView.layoutIfNeeded()
-            bottomViewHeight = quizControlView.bounds.height + (margin * 2)
-        } else {
-            dictationStackView.snp.makeConstraints { make in
-                make.left.right.equalTo(safeArea).inset(margin + 4)
-                self.bottomConstraint = make.bottom.equalTo(safeArea).inset(margin).constraint
-            }
-            dictationStackView.layoutIfNeeded()
-            bottomViewHeight = dictationStackView.bounds.height + (margin * 2)
-        }
-        
-        cardViewBottomConstraint?.update(inset: bottomViewHeight)
-    }
-    
     func textFieldSetup() {
-        dictationTextField.delegate = self
-        dictationTextField.becomeFirstResponder()
+        quizMainView.dictationTextField.delegate = self
+        quizMainView.dictationTextField.becomeFirstResponder()
     }
     
     func keyboardReturn() {
         if viewModel.isMatch {
-            cardMoveAnimation(moveView: frontCardView, isMemorize: true)
+            cardMoveAnimation(moveView: quizMainView.frontCardView, isMemorize: true)
         }
+    }
+    
+    func setupCardView() {
+        if viewModel.isComplete {
+            navigationPushQuizCompletePage()
+            return
+        }
+        quizMainView.setupCardView()
     }
     
     func navigationPushQuizCompletePage() {
@@ -237,10 +116,11 @@ private extension QuizViewController {
         vc.popCompletion = { [weak self] in
             guard let self else { return }
             self.viewModel.reQuizWordSetting()
-            self.backgroundCardView.isHidden = false
-            self.cardViewSetup()
+            self.quizMainView.backgroundCardView.isHidden = false
+            self.setupCardView()
             if self.viewModel.isTestType {
-                timerReset()
+                timer.invalidate()
+                quizMainView.timerReset()
             }
         }
         navigationController?.pushViewController(vc, animated: false)
@@ -258,21 +138,10 @@ private extension QuizViewController {
                 cardMoveAnimation(moveView: moveView, isMemorize: true)
             } else {
                 UIView.animate(withDuration: 0.3) {
-                    moveView.center = self.backgroundCardView.center
+                    moveView.center = self.quizMainView.backgroundCardView.center
                 }
             }
         }
-    }
-    
-    private func textFieldInit() {
-        dictationTextField.text = ""
-    }
-    
-    private func timerReset() {
-        timer.invalidate()
-        timerSettingView.isHidden = false
-        countDownLabel.text = "00시00분00초"
-        countDownLabel.textColor = .black
     }
     
     private func cardMoveAnimation(moveView: UIView, isMemorize: Bool) {
@@ -285,8 +154,8 @@ private extension QuizViewController {
             self.viewModel.repository.update()
             self.viewModel.quizResultWords.append(word)
             self.viewModel.next()
-            self.textFieldInit()
-            self.cardViewSetup()
+            self.quizMainView.textFieldInit()
+            self.setupCardView()
         }
     }
     
@@ -295,11 +164,11 @@ private extension QuizViewController {
         let keyboardFrame:NSValue = userInfo.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as! NSValue
         let keyboardRectangle = keyboardFrame.cgRectValue
         let keyboardHeight = keyboardRectangle.height
-        self.bottomConstraint?.update(inset: keyboardHeight - 20)
+        self.quizMainView.bottomConstraint?.update(inset: keyboardHeight - 20)
     }
 
     @objc func keyBoardWillHide(notification: NSNotification) {
-        self.bottomConstraint?.update(inset: margin)
+        self.quizMainView.bottomConstraint?.update(inset: margin)
     }
     
     func bind() {
@@ -315,18 +184,18 @@ private extension QuizViewController {
         }
         viewModel.hiddenQuizObservable.bind { [weak self] isHidden in
             guard let self else { return }
-            if self.viewModel.isMeanDictation { self.frontCardView.koreaLabel.alpha = isHidden ? 0.0 : 1 }
-            else if self.viewModel.isWordDictation { self.frontCardView.englishLabel.alpha = isHidden ? 0.0 : 1 }
+            if self.viewModel.isMeanDictation { self.quizMainView.frontCardView.koreaLabel.alpha = isHidden ? 0.0 : 1 }
+            else if self.viewModel.isWordDictation { self.quizMainView.frontCardView.englishLabel.alpha = isHidden ? 0.0 : 1 }
         }
         viewModel.nextHandler = { [weak self] isMemorize in
             guard let self else { return }
-            self.cardMoveAnimation(moveView: frontCardView, isMemorize: isMemorize)
+            self.cardMoveAnimation(moveView: quizMainView.frontCardView, isMemorize: isMemorize)
         }
         viewModel.reQuizHandler = { [weak self] in
             guard let self else { return }
-            self.backgroundCardView.isHidden = false
+            self.quizMainView.backgroundCardView.isHidden = false
             self.viewModel.reQuizWordSetting()
-            self.cardViewSetup()
+            self.setupCardView()
         }
         viewModel.dismissHandler = { [weak self] in
             self?.dismiss(animated: true)
@@ -346,7 +215,7 @@ private extension QuizViewController {
                 return
             }
             
-            self?.countDownTime = remainSeconds
+            self?.quizMainView.countDownTime = remainSeconds
         })
     }
     
@@ -359,14 +228,12 @@ private extension QuizViewController {
 extension QuizViewController: UITextFieldDelegate {
     func textFieldDidChangeSelection(_ textField: UITextField) {
         viewModel.dictationText = textField.text
-        frontCardView.koreaLabel.text = viewModel.currentKoreaLabelText
-        frontCardView.englishLabel.text = viewModel.currentEnglishLabelText
-        
+        quizMainView.frontCardView.koreaLabel.text = viewModel.currentKoreaLabelText
+        quizMainView.frontCardView.englishLabel.text = viewModel.currentEnglishLabelText
     }
     
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         keyboardReturn()
         return true
     }
-    
 }
